Snap fixes from review; snap docs
- TableFitter: Fit.iou is Float (compile error).
- Y pan clamped in ApplyTablePose and in the fitter's current pose, as
  UpdateStateUseCase clamps it; otherwise the pull drove the stored offset
  past the drawn one without bound.
- Lock applies the fit pose in the reducer for the same event instead of
  queuing ApplyTablePose, which a fast lock result could have skipped.
- Camera-on seed skipped if the user moved the table during the GPS fix, or
  in beginner-locked view.
- Snap/memory docs (re-added without the removed top-down parts).

+++ app/src/main/java/com/hereliesaz/cuedetat/data/TableFitter.kt
@@ -113,7 +113,7 @@ class TableFitter {
             TableOrientationLearner.normalize180(best.params[2].toFloat()), exp(best.params[3]).toFloat(),
         )
         val quad = viewQuad(pose, a, corners)
-        return Fit(pose, iou(pose, a, viewToFrame, corners), quad)
+        return Fit(pose, iou(pose, a, viewToFrame, corners).toFloat(), quad)
     }
 
     private fun worldMatrix(p: Pose) = Matrix().apply {
